phase 1/resume-API/flask_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
python-jose==3.3.0
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
SQLAlchemy-Utils==0.41.1
PyMuPDF==1.23.8
python-docx==1.1.0
spacy==3.7.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
marshmallow==3.20.1
python-json-logger==2.0.7
redis==5.0.1
ratelimit==2.2.1
